Move more preprocessing functions into common/model_utils
